Split git_test To Use Git Server
Enable e2e tests to use git server
